+++ circuits/tests/transaction.rs
@@ -3,7 +3,13 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
-use rand_core::{CryptoRng, OsRng, RngCore};
+use rand::rngs::StdRng;
+use rand::SeedableRng;
+use rand::{CryptoRng, RngCore};
+
+use dusk_jubjub::JubJubScalar;
+use phoenix_circuits::transaction::{TxCircuit, TxInputNote, TxOutputNote};
+use phoenix_core::{Note, PublicKey, SecretKey};
 
 use dusk_plonk::prelude::*;
 use ff::Field;
@@ -24,43 +30,44 @@ const CAPACITY: usize = 17; // capacity required for the setup
 const HEIGHT: usize = 17;
 
 struct TestingParameters {
-    sk: SecretKey,
     pp: PublicParameters,
     tx_input_notes: [TxInputNote<HEIGHT>; 4],
     payload_hash: BlsScalar,
     root: BlsScalar,
-    crossover: u64,
+    deposit: u64,
     max_fee: u64,
     rp: RecipientParameters,
 }
 
 lazy_static! {
     static ref TP: TestingParameters = {
-        let pp = PublicParameters::setup(1 << CAPACITY, &mut OsRng).unwrap();
-        let sk = SecretKey::random(&mut OsRng);
+        let mut rng = StdRng::seed_from_u64(0xc0b);
+
+        let pp = PublicParameters::setup(1 << CAPACITY, &mut rng).unwrap();
+        let sk = SecretKey::random(&mut rng);
 
         let mut tree = Tree::<(), HEIGHT>::new();
         let payload_hash = BlsScalar::from(1234u64);
 
         // create and insert into the tree 4 testing tx input notes
         let tx_input_notes =
-            create_test_tx_input_notes::<4>(&mut tree, &sk, payload_hash, &mut OsRng);
+            create_test_tx_input_notes::<4>(&mut rng, &mut tree, &sk, payload_hash);
 
         // retrieve the root from the tree after inserting the notes
         let root = tree.root().hash;
 
-        let crossover = 5;
+        let deposit = 5;
         let max_fee = 5;
 
         // Compute the tx output note public keys using
         // the receiver public key
         let pk_receiver = PublicKey::from(&sk);
 
-        let r = JubJubScalar::random(&mut OsRng);
+        let r = JubJubScalar::random(&mut rng);
         let sa = pk_receiver.gen_stealth_address(&r);
         let note_pk_1 = sa.note_pk();
 
-        let r = JubJubScalar::random(&mut OsRng);
+        let r = JubJubScalar::random(&mut rng);
         let sa = pk_receiver.gen_stealth_address(&r);
         let note_pk_2 = sa.note_pk();
 
@@ -75,24 +82,24 @@ lazy_static! {
         let recipient_pk = PublicKey::from(&sk);
 
         #[allow(non_snake_case)]
-        let r_A_1 = JubJubScalar::random(&mut OsRng);
+        let r_A_1 = JubJubScalar::random(&mut rng);
         #[allow(non_snake_case)]
         let (A_enc_1_c1, A_enc_1_c2) =
             elgamal::encrypt(note_pk_1.as_ref(), &recipient_pk.A(), &r_A_1);
 
         #[allow(non_snake_case)]
-        let r_B_1 = JubJubScalar::random(&mut OsRng);
+        let r_B_1 = JubJubScalar::random(&mut rng);
         #[allow(non_snake_case)]
         let (B_enc_1_c1, B_enc_1_c2) =
             elgamal::encrypt(note_pk_1.as_ref(), &recipient_pk.B(), &r_B_1);
         #[allow(non_snake_case)]
-        let r_A_2 = JubJubScalar::random(&mut OsRng);
+        let r_A_2 = JubJubScalar::random(&mut rng);
         #[allow(non_snake_case)]
         let (A_enc_2_c1, A_enc_2_c2) =
             elgamal::encrypt(note_pk_2.as_ref(), &recipient_pk.A(), &r_A_2);
 
         #[allow(non_snake_case)]
-        let r_B_2 = JubJubScalar::random(&mut OsRng);
+        let r_B_2 = JubJubScalar::random(&mut rng);
         #[allow(non_snake_case)]
         let (B_enc_2_c1, B_enc_2_c2) =
             elgamal::encrypt(note_pk_2.as_ref(), &recipient_pk.B(), &r_B_2);
@@ -110,11 +117,11 @@ lazy_static! {
         // Sign the payload hash using both 'a' and 'b'
         let schnorr_sk_a = SchnorrSecretKey::from(sk.a());
         #[allow(non_snake_case)]
-        let sig_A = schnorr_sk_a.sign(&mut OsRng, payload_hash);
+        let sig_A = schnorr_sk_a.sign(&mut rng, payload_hash);
 
         let schnorr_sk_b = SchnorrSecretKey::from(sk.b());
         #[allow(non_snake_case)]
-        let sig_B = schnorr_sk_b.sign(&mut OsRng, payload_hash);
+        let sig_B = schnorr_sk_b.sign(&mut rng, payload_hash);
 
         let sig_vec = [sig_A, sig_B];
 
@@ -125,11 +132,11 @@ lazy_static! {
 }
 
 fn create_and_insert_test_note(
+    rng: &mut (impl RngCore + CryptoRng),
     tree: &mut Tree<(), HEIGHT>,
     pk: &PublicKey,
     pos: u64,
     value: u64,
-    rng: &mut (impl RngCore + CryptoRng),
 ) -> Note {
     let mut note = Note::transparent(rng, pk, value);
     note.set_pos(pos);
@@ -144,21 +151,21 @@ fn create_and_insert_test_note(
 }
 
 fn create_test_tx_input_notes<const I: usize>(
+    rng: &mut (impl RngCore + CryptoRng),
     tree: &mut Tree<(), HEIGHT>,
     sk: &SecretKey,
     payload_hash: BlsScalar,
-    rng: &mut (impl RngCore + CryptoRng),
 ) -> [TxInputNote<HEIGHT>; I] {
     let pk = PublicKey::from(sk);
 
     let mut notes = Vec::new();
     for i in 0..I {
         notes.push(create_and_insert_test_note(
+            rng,
             tree,
             &pk,
             i.try_into().unwrap(),
             25,
-            rng,
         ));
     }
 
@@ -166,11 +173,11 @@ fn create_test_tx_input_notes<const I: usize>(
     for i in 0..I {
         let merkle_opening = tree.opening(*notes[i].pos()).expect("Tree read.");
         let input_note = TxInputNote::new(
+            rng,
             &notes[i],
             merkle_opening,
             &sk,
             &payload_hash,
-            rng,
         )
         .expect("Note created properly.");
 
@@ -180,18 +187,17 @@ fn create_test_tx_input_notes<const I: usize>(
     input_notes.try_into().unwrap()
 }
 
-fn create_test_tx_output_note(
-    sk: &SecretKey,
-    value: u64,
-    rng: &mut (impl RngCore + CryptoRng),
-) -> TxOutputNote {
-    let note = Note::transparent(rng, &PublicKey::from(sk), value);
-    TxOutputNote::new(&note, &ViewKey::from(&TP.sk))
-        .expect("Note created properly.")
+// we don't care if the test output notes are spendable
+fn create_test_tx_output_note(value: u64) -> TxOutputNote {
+    let blinding_factor = JubJubScalar::from(42u64);
+
+    TxOutputNote::new(value, blinding_factor)
 }
 
 #[test]
 fn test_transfer_circuit_1_2() {
+    let mut rng = StdRng::seed_from_u64(0xc0b);
+
     let (prover, verifier) =
         Compiler::compile::<TxCircuit<HEIGHT, 1>>(&TP.pp, LABEL)
             .expect("failed to compile circuit");
@@ -200,19 +206,19 @@ fn test_transfer_circuit_1_2() {
 
     // create 2 testing tx output notes
     let tx_output_notes = [
-        create_test_tx_output_note(&TP.sk, 10, &mut OsRng),
-        create_test_tx_output_note(&TP.sk, 5, &mut OsRng),
+        create_test_tx_output_note(10),
+        create_test_tx_output_note(5),
     ];
 
     let (proof, public_inputs) = prover
         .prove(
-            &mut OsRng,
+            &mut rng,
             &TxCircuit::new(
                 input_notes,
                 tx_output_notes,
                 TP.payload_hash,
                 TP.root,
-                TP.crossover,
+                TP.deposit,
                 TP.max_fee,
                 TP.rp,
             ),
@@ -226,6 +232,8 @@ fn test_transfer_circuit_1_2() {
 
 #[test]
 fn test_transfer_circuit_2_2() {
+    let mut rng = StdRng::seed_from_u64(0xc0b);
+
     let (prover, verifier) =
         Compiler::compile::<TxCircuit<HEIGHT, 2>>(&TP.pp, LABEL)
             .expect("failed to compile circuit");
@@ -235,19 +243,19 @@ fn test_transfer_circuit_2_2() {
 
     // create 2 testing tx output notes
     let tx_output_notes = [
-        create_test_tx_output_note(&TP.sk, 35, &mut OsRng),
-        create_test_tx_output_note(&TP.sk, 5, &mut OsRng),
+        create_test_tx_output_note(35),
+        create_test_tx_output_note(5),
     ];
 
     let (proof, public_inputs) = prover
         .prove(
-            &mut OsRng,
+            &mut rng,
             &TxCircuit::new(
                 input_notes,
                 tx_output_notes,
                 TP.payload_hash,
                 TP.root,
-                TP.crossover,
+                TP.deposit,
                 TP.max_fee,
                 TP.rp,
             ),
@@ -261,6 +269,8 @@ fn test_transfer_circuit_2_2() {
 
 #[test]
 fn test_transfer_circuit_3_2() {
+    let mut rng = StdRng::seed_from_u64(0xc0b);
+
     let (prover, verifier) =
         Compiler::compile::<TxCircuit<HEIGHT, 3>>(&TP.pp, LABEL)
             .expect("failed to compile circuit");
@@ -273,19 +283,19 @@ fn test_transfer_circuit_3_2() {
 
     // create 2 testing tx output notes
     let tx_output_notes = [
-        create_test_tx_output_note(&TP.sk, 35, &mut OsRng),
-        create_test_tx_output_note(&TP.sk, 30, &mut OsRng),
+        create_test_tx_output_note(35),
+        create_test_tx_output_note(30),
     ];
 
     let (proof, public_inputs) = prover
         .prove(
-            &mut OsRng,
+            &mut rng,
             &TxCircuit::new(
                 input_notes,
                 tx_output_notes,
                 TP.payload_hash,
                 TP.root,
-                TP.crossover,
+                TP.deposit,
                 TP.max_fee,
                 TP.rp,
             ),
@@ -299,25 +309,27 @@ fn test_transfer_circuit_3_2() {
 
 #[test]
 fn test_transfer_circuit_4_2() {
+    let mut rng = StdRng::seed_from_u64(0xc0b);
+
     let (prover, verifier) =
         Compiler::compile::<TxCircuit<HEIGHT, 4>>(&TP.pp, LABEL)
             .expect("failed to compile circuit");
 
     // create 2 testing tx output notes
     let tx_output_notes = [
-        create_test_tx_output_note(&TP.sk, 60, &mut OsRng),
-        create_test_tx_output_note(&TP.sk, 30, &mut OsRng),
+        create_test_tx_output_note(60),
+        create_test_tx_output_note(30),
     ];
 
     let (proof, public_inputs) = prover
         .prove(
-            &mut OsRng,
+            &mut rng,
             &TxCircuit::new(
                 TP.tx_input_notes.clone(),
                 tx_output_notes,
                 TP.payload_hash,
                 TP.root,
-                TP.crossover,
+                TP.deposit,
                 TP.max_fee,
                 TP.rp,
             ),