+++ core/tests/transaction.rs
@@ -10,7 +10,7 @@ use dusk_bls12_381::BlsScalar;
 use dusk_jubjub::JubJubScalar;
 use ff::Field;
 use phoenix_core::{Error, Note, PublicKey, SecretKey, TxSkeleton};
-use rand_core::OsRng;
+use rand::rngs::OsRng;
 
 #[test]
 fn transaction_parse() -> Result<(), Error> {
@@ -27,14 +27,14 @@ fn transaction_parse() -> Result<(), Error> {
     let nullifiers = vec![BlsScalar::from(456), BlsScalar::from(789)];
     let outputs = [note.clone(), note];
     let tx_max_fee = 0;
-    let crossover = 0;
+    let deposit = 0;
 
     let tx_skeleton = TxSkeleton {
         root,
         nullifiers,
         outputs,
         tx_max_fee,
-        crossover,
+        deposit,
     };
     let bytes_of_transaction = tx_skeleton.to_var_bytes();
     assert_eq!(