+++ circuits/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Remove `ViewKey` from `TxOutputNote::new()` parameters [#191]
+- Make `rng` the first param in `TxInputNote::new` [#189]
+- Rename `crossover` to `deposit` [#190]
+
 ## [0.1.0] - 2024-05-22
 
 ### Added
@@ -23,6 +29,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update `poseidon-merkle` to v0.6 [#179]
 
 <!-- ISSUES -->
+[#191]: https://github.com/dusk-network/phoenix/issues/191
+[#190]: https://github.com/dusk-network/phoenix/issues/190
+[#189]: https://github.com/dusk-network/phoenix/issues/189
 [#179]: https://github.com/dusk-network/phoenix/issues/179
 [#177]: https://github.com/dusk-network/phoenix/issues/177
 [#171]: https://github.com/dusk-network/phoenix/issues/171