+++ core/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Rename `crossover` to `deposit` [#190]
+
+### Removed
+
+- Remove `"getrandom"` feature from `aes-gcm` dependency [#195]
+
 ## [0.28.1] - 2024-05-23
 
 ### Changed
@@ -317,6 +325,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Canonical implementation shielded by feature.
 
 <!-- ISSUES -->
+[#195]: https://github.com/dusk-network/phoenix/issues/195
+[#190]: https://github.com/dusk-network/phoenix/issues/190
 [#183]: https://github.com/dusk-network/phoenix/issues/183
 [#179]: https://github.com/dusk-network/phoenix/issues/179
 [#175]: https://github.com/dusk-network/phoenix/issues/175